fix: move python script to own file and handle conditions where base container does not have python installed
